--- qrsignup/src/adminpage/Admin.css
@@ -1,47 +1,0 @@
-.AdminCenter {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.AdminCard {
-  width: 442px;
-  height: 500px;
-  border-radius: 10px;
-  background: #53536c;
-  box-shadow: 0px 12px 14px rgba(0, 0, 0, 0.25);
-  text-align: center;
-}
-
-.AdminCard .Compo {
-  height: 25%;
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  align-items: center;
-  justify-content: center;
-}
-
-.AdminCard p {
-  width:100%;
-  text-align: center;
-  color: white;
-}
-
-.AdminCard input {
-  width: 80%;
-  border-radius: 10px;
-  height: 30px;
-}
-
-.AdminCard button {
-  width: 80%;
-  border-radius: 10px;
-  height: 40px;
-  cursor: pointer;
-}
