--- qrsignup/src/main/Main.css
@@ -1,164 +1,0 @@
-.Master {
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  align-items: center;
-  justify-content: center;
-}
-
-.Logo {
-  position: absolute;
-  top: 15px;
-  right: 15px;
-  height: 35.12;
-  width: 117.56px;
-}
-
-.Logo img {
-  width: 100px;
-  height: auto;
-  object-fit: cover;
-}
-
-.Brand {
-  display: flex;
-  flex-direction: column;
-  width: 70%;
-  height: 30%;
-  justify-content: center;
-  align-items: center;
-}
-
-.Brand h1 {
-  font-size: 36px;
-  font-weight: 600;
-}
-
-.Navigation {
-  display: flex;
-  height: 50%;
-  justify-content: space-around;
-  width: 80%;
-}
-
-.Navigation a {
-  color: #000;
-  text-decoration: none;
-}
-
-.Navigation a:visited {
-  color: #000;
-  text-decoration: none;
-}
-
-.LeftCard {
-  display: flex;
-  flex-direction: column;
-  width: 300px;
-  height: 400px;
-  border-radius: 10px;
-  background: #53536c;
-  box-shadow: 10px 10px 30px #47475c;
-  justify-content: center;
-  align-items: center;
-}
-
-.LeftCard h3 {
-  color: white;
-}
-
-.RightCard {
-  display: flex;
-  flex-direction: column;
-  width: 300px;
-  height: 400px;
-  border-radius: 10px;
-  background: #53536c;
-  box-shadow: 10px 10px 30px #47475c;
-  justify-content: center;
-  align-items: center;
-}
-
-.RightCard h3 {
-  color: white;
-}
-
-.TitlePart {
-  height: 20%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 24px;
-  color: white;
-}
-
-.ImagePart {
-  height: 40%;
-  width: 60%;
-  margin: 5px;
-  border-radius: 15px;
-  background: #53536c;
-  box-shadow: 5px 5px 15px #47475c, -5px -5px 15px #5f5f7c;
-  overflow: hidden;
-}
-
-.ImagePart img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-@media all and (min-width: 200px) and (max-width: 900px) {
-  /*스타일입력*/
-  .Navigation {
-    display: flex;
-    flex-direction: column;
-    height: 90%;
-    justify-content: space-around;
-    align-items: center;
-    width: 100%;
-  }
-
-  .LeftCard {
-    display: flex;
-    flex-direction: column;
-    width: 300px;
-    height: 300px;
-    border-radius: 10px;
-    background: #53536c;
-    box-shadow: 10px 10px 30px #47475c;
-    justify-content: center;
-    align-items: center;
-    margin-bottom: 50px;
-  }
-
-  .RightCard {
-    display: flex;
-    flex-direction: column;
-    width: 300px;
-    height: 300px;
-    border-radius: 10px;
-    background: #53536c;
-    box-shadow: 10px 10px 30px #47475c;
-    justify-content: center;
-    align-items: center;
-    margin-bottom: 50px;
-  }
-}
-
-/* 테블릿 세로 (해상도 768px ~ 1023px)*/
-@media all and (min-width: 768px) and (max-width: 1023px) {
-  /*스타일입력*/
-}
-
-/* 모바일 가로, 테블릿 세로 (해상도 480px ~ 767px)*/
-@media all and (min-width: 480px) and (max-width: 767px) {
-  /*스타일입력*/
-}
-
-/* 모바일 가로, 테블릿 세로 (해상도 ~ 479px)*/
-@media all and (max-width: 479px) {
-  /*스타일입력*/
-}
