+++ qrsignup/src/screens/register/RegisterContainer.js
@@ -1,0 +1,7 @@
+import RegisterPresenter from "./RegisterPresenter";
+
+const RegisterContainer = () => {
+  return <RegisterPresenter />;
+};
+
+export default RegisterContainer;
