+++ qrsignup/src/screens/register/RegisterPresenter.js
@@ -1,0 +1,8 @@
+import shop from "../../img/shop.png";
+import logo from "../../img/kt.png";
+
+const RegisterPresenter = () => {
+  return <div>dd</div>;
+};
+
+export default RegisterPresenter;
