+++ qrsignup/src/screens/register/index.js
@@ -1,0 +1,4 @@
+import RegisterContainer from "./RegisterContainer";
+import "../statics/Register.css";
+
+export default RegisterContainer;
