+++ qrsignup/src/screens/wifi/index.js
@@ -1,0 +1,4 @@
+import WifiContainer from "./WifiContainer";
+import "../statics/Wifi.scss";
+
+export default WifiContainer;
