--- qrsignup/src/wifi/Wifi.scss
@@ -1,326 +1,0 @@
-:root {
-  --input-color: #000;
-  --input-border: #000;
-  --input-background: #fff;
-  --input-placeholder: #CBD1DC;
-
-  --input-border-focus: #B5ACAD;
-
-  --group-color: var(--input-color);
-  --group-border: var(--input-border);
-  --group-background: #FFEAE3;
-
-  --group-color-focus: #fff;
-  --group-border-focus: var(--input-border-focus);
-  --group-background-focus: #B5ACAD;
-
-}
-
-.Logo1 {
-  position: absolute;
-  top:15px;
-  right: 15px;
-  height: 35.12;
-  width: 117.56px;
-  
-}
-
-.Logo1 img {
-  width: 100px;
-  height: auto;
-  object-fit: cover;
-}
-
-.form-field {
-  display: block;
-  width: 100%;
-  padding: 8px 16px;
-  line-height: 25px;
-  font-size: 14px;
-  font-weight: 500;
-  font-family: inherit;
-  border-radius: 6px;
-  -webkit-appearance: none;
-  color: var(--input-color);
-  border: 2px solid var(--input-border);
-  background: var(--input-background);
-  transition: border .3s ease;
-  &::placeholder {
-      color: var(--input-placeholder);
-  }
-  &:focus {
-      outline: none;
-      border-color: var(--input-border-focus);
-  }
-}
-
-.form-group {
-  position: relative;
-  display: flex;
-  width: 60%;
-  & > span,
-  .form-field {
-      white-space: nowrap;
-      display: block;
-      &:not(:first-child):not(:last-child) {
-          border-radius: 0;
-      }
-      &:first-child {
-          border-radius: 6px 0 0 6px;
-      }
-      &:last-child {
-          border-radius: 0 6px 6px 0;
-      }
-      &:not(:first-child) {
-          margin-left: -1px;
-      }
-  }
-  .form-field {
-      position: relative;
-      z-index: 1;
-      flex: 1 1 auto;
-      width: 1%;
-      margin-top: 0;
-      margin-bottom: 0;
-  }
-  & > span {
-      text-align: center;
-      padding: 8px 12px;
-      font-size: 14px;
-      line-height: 25px;
-      color: var(--group-color);
-      background: var(--group-background);
-      border: 2px solid var(--group-border);
-      transition: background .3s ease, border .3s ease, color .3s ease;
-  }
-  &:focus-within {
-      & > span {
-          color: var(--group-color-focus);
-          background: var(--group-background-focus);
-          border-color: var(--group-border-focus);
-      }
-  }
-}
-
-
-.Di {
-  height: 150%;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: #DED8DA;
-}
-
-.Buttons {
-  width: 300px;
-  display: flex;
-  justify-content: space-between;
-  padding-top: 20px;
-  padding-bottom: 20px;
-}
-
-.Back {
-  height: 1124px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  background-color: #DED8DA;
-}
-
-.Di2 {
-  height: 1024px;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: #DED8DA;
-}
-
-.CardComp {
-  width: 500px;
-  height: 800px;
-  border-radius: 10px;
-  border: 2px solid black;
-  background-color: white;
-  box-shadow: 6px 12px 14px rgba(0, 0, 0, 0.25);
-  display: flex;
-  flex-direction: column;
-}
-
-.Pp1 {
-  height: 15%;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
-
-.Pp2 {
-  height: 20%;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
-
-.Pp4 {
-  height: 20%;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
-
-.Pp5 {
-  height: 5%;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
-
-.Center {
-  width: 719px;
-  height:1024px;
-  overflow: hidden;
-  position: relative;
-  color: white;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
-
-.Center img {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  z-index: 2;
-}
-
-.P1 {
-  height: 22.16%;
-  width: 100%;
-}
-
-.P2 {
-  height: 7.72%;
-  width: 100%;
-  z-index: 3;
-}
-
-.P3 {
-  height: 23.05%;
-  width: 100%;
-}
-
-.P4 {
-  height: 32.22%;
-  width: 100%;
-  z-index: 3;
-}
-
-.P5 {
-  height: 14.84%;
-  width: 100%;
-  z-index: 3;
-  font-size: 24px;
-}
-
-/* CSS */
-.button-74 {
-  background-color: #fbeee0;
-  border: 2px solid #422800;
-  border-radius: 10px;
-  box-shadow: #422800 4px 4px 0 0;
-  color: #422800;
-  cursor: pointer;
-  display: inline-block;
-  font-weight: 600;
-  font-size: 18px;
-  padding: 0 18px;
-  line-height: 50px;
-  text-align: center;
-  text-decoration: none;
-  user-select: none;
-  -webkit-user-select: none;
-  touch-action: manipulation;
-}
-
-.button-74:hover {
-  background-color: #fff;
-}
-
-.button-74:active {
-  box-shadow: #422800 2px 2px 0 0;
-  transform: translate(2px, 2px);
-}
-
-/* CSS */
-.button-16 {
-  background-color: #f8f9fa;
-  border: 1px solid #f8f9fa;
-  border-radius: 4px;
-  color: #3c4043;
-  cursor: pointer;
-  font-family: arial,sans-serif;
-  font-size: 14px;
-  height: 36px;
-  line-height: 27px;
-  min-width: 54px;
-  padding: 0 16px;
-  text-align: center;
-  user-select: none;
-  -webkit-user-select: none;
-  touch-action: manipulation;
-  white-space: pre;
-}
-
-.button-16:hover {
-  border-color: #dadce0;
-  box-shadow: rgba(0, 0, 0, .1) 0 1px 1px;
-  color: #202124;
-}
-
-.button-16:focus {
-  border-color: #4285f4;
-  outline: none;
-}
-
-
-@media all and (min-width:768px) {
-  /*스타일입력*/
-  .button-74 {
-    min-width: 120px;
-    padding: 0 25px;
-  }
-} 
-
-/* 모바일 가로, 테블릿 세로 (해상도 480px ~ 767px)*/ 
-@media all and (min-width:480px) and (max-width:767px) {
-  /*스타일입력*/
-  .Center {
-    width: 100%;
-    height:1024px;
-    overflow: hidden;
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-  }
-} 
-
-/* 모바일 가로, 테블릿 세로 (해상도 ~ 479px)*/ 
-@media all and (max-width:479px) {
-  /*스타일입력*/
-}
